requirements.txt
aiohttp==3.8.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.1.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.2
asyncssh==2.12.0
atpublic==3.1.1
attrs==22.1.0
billiard==3.6.4.0
black==22.10.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
commonmark==0.9.1
configobj==5.0.6
cryptography==38.0.3
dictdiffer==0.9.0
diskcache==5.4.0
distro==1.8.0
dpath==2.0.6
dulwich==0.20.50
dvc==2.34.0
dvc-data==0.25.3
dvc-http==2.27.2
dvc-objects==0.12.2
dvc-render==0.0.12
dvc-task==0.1.4
dvclive==1.0.1
filelock==3.8.0
flatten-dict==0.4.2
flufl.lock==7.1.1
frozenlist==1.3.3
fsspec==2022.11.0
funcy==1.17
future==0.18.2
gitdb==4.0.9
GitPython==3.1.29
grandalf==0.6
hydra-core==1.2.0
idna==3.4
iterative-telemetry==0.0.5
joblib==1.2.0
kombu==5.2.4
lightgbm==3.3.3
multidict==6.0.2
mypy-extensions==0.4.3
nanotime==0.5.2
networkx==2.8.8
numpy==1.23.4
omegaconf==2.2.3
packaging==21.3
pandas==1.5.1
pathspec==0.9.0
pip==22.3
platformdirs==2.5.2
prompt-toolkit==3.0.32
psutil==5.9.4
pycparser==2.21
pydot==1.4.2
pygit2==1.11.1
Pygments==2.13.0
pygtrie==2.5.0
pyparsing==3.0.9
python-box==6.1.0
python-dateutil==2.8.2
pytz==2022.5
PyYAML==6.0
requests==2.28.1
rich==12.6.0
ruamel.yaml==0.17.21
scikit-learn==1.1.3
scipy==1.9.3
scmrepo==0.1.3
setuptools==65.5.0
shortuuid==1.0.10
shtab==1.5.7
six==1.16.0
smmap==5.0.0
tabulate==0.9.0
threadpoolctl==3.1.0
tomlkit==0.11.6
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.12
vine==5.0.0
voluptuous==0.13.1
wcwidth==0.2.5
wheel==0.37.1
yarl==1.8.1
zc.lockfile==2.0
